% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{add_meta}
\alias{add_meta}
\title{Join user-defined metadata from `uncleR::get_meta` function to imported Uncle data}
\usage{
add_meta(data, meta)
}
\arguments{
\item{data}{a dataframe to assign metadata to}

\item{meta}{a named list of tibbles containing metadata for experiments imported from `uncleR::get_meta` function
defaults to c("well")}
}
\value{
a dataframe with metadata added as new variables, matched according to well
}
\description{
\code{add_meta}
}
