% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{assign_metadata}
\alias{assign_metadata}
\title{Assign user-defined metadata to imported DLS or Tm/Tagg SLS summary dataframe}
\usage{
assign_metadata(data, metaCSV_path, .by = c("well"))
}
\arguments{
\item{metaCSV_path}{a path (relative to current working directory) to a .csv file containing metadata variables corresponding to a well (e.g. A1, E4..)}

\item{.by}{a character vector of variable name(s) by which to join metadata;
defaults to c("well")}
}
\value{
a dataframe with metadata added as new variables, matched according to well
}
\description{
\code{assign_metadata}
}
