% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{consolidate_experiments}
\alias{consolidate_experiments}
\title{Consolidates Uncle Experiments}
\usage{
consolidate_experiments(
  prot_dir,
  join_vars = NULL,
  legacy = FALSE,
  SLSheader = TRUE,
  DLSheader = FALSE
)
}
\arguments{
\item{prot_dir}{a character string identical to subdirectory name of working directory
where the experiment "Exports" folder is located for the protein being analyzed}

\item{legacy}{a boolean value, TRUE if individual SLS/DLS spectra files were exported,
FALSE if SLS/DLS "bundle" files were exported; default is FALSE}

\item{SLSheader}{sets header argument for \code{uncleR::import_SLSsum}; defaults to TRUE}

\item{DLSheader}{sets header argument for \code{uncleR::import_DLSsum}; defaults to FALSE}
}
\value{
a named list of dataframes containing the consolidated data for each experiment
}
\description{
This function consolidates experiments for a protein based on directory hierarchy.
The hierarchy should be as follows from within the working directory where these
functions will be executed:
}
\details{
<wd>/
   <prot_dir>/
       "Exports"/
           "General Screen"
           "pH Screen"
           ...

The \code{prot_dir} argument should match the name of the protein subdirectory within
the working directory. This is deliberate as it forces the user to make
a coherent choice about which data to process. Exported summary and spectra
should be contained within named subdirectories of a directory named "Exports"
located within the \code{prot_dir} directory.

\code{consolidate_experiments}
}
