% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLS.R
\name{import_DLSsum}
\alias{import_DLSsum}
\title{Import UNcle Tm/Tagg DLS summary into R}
\usage{
import_DLSsum(directory_path, pattern = ".*\\\\.xlsx", sheet = NULL,
  temp_cutoff = 100, header = FALSE, combine = TRUE)
}
\arguments{
\item{directory_path}{a path to a directory containing the exported .xlsx files}

\item{pattern}{a regex pattern for further selecing files in the directory,
defaults to reading all .xlsx files present}

\item{sheet}{character string to specify sheet if multi-sheet workbook is exported}

\item{temp_cutoff}{numeric value, excluding all DLS data obtained at temperatures above this value, default is 100 (°C)}

\item{header}{if TRUE skips first 4 rows of .xlsx file to remove UNcle header, default is FALSE}

\item{combine}{if TRUE, returns all imported data merged into one unified dataframe with an "origin" column listing the original file path,
FALSE will return a list of dataframes; default is TRUE}
}
\value{
a named (with filename) list of dataframes or a single merged dataframe
}
\description{
\code{import_DLSsum}
}
