% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_dynamicBundle}
\alias{import_dynamicBundle}
\title{Import dynamic UNcle spectra (DLS) into R}
\usage{
import_dynamicBundle(
  directory_path,
  pattern = "DLS Bundle",
  skip = 2,
  combine = TRUE
)
}
\arguments{
\item{directory_path}{a path to a directory containing the exported .xlsx files}

\item{pattern}{a regex for narrowing selection of files in the `directory_path`;
defaults to "DLS Bundle"}

\item{skip}{number of rows of embedded run metadata in the .xlsx file to remove; default is 2}

\item{combine}{if TRUE, returns all imported data merged into one unified dataframe with an "origin" column listing the original file path,
FALSE will return a list of dataframes; default is TRUE}
}
\value{
a named (with filename) list of dataframes or a single merged dataframe
}
\description{
\code{import_dynamicBundle}
}
\details{
Note: This import function will only import DLS spectra performed at 
the beginning of the temperature ramp, i.e. the lowest temperature.
}
