% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLUOR.R
\name{import_FLUORspec}
\alias{import_FLUORspec}
\title{Import Uncle fluorescence spectra into R}
\usage{
import_FLUORspec(
  directory_path,
  pattern = "Tm Spec",
  header = TRUE,
  combine = TRUE
)
}
\arguments{
\item{directory_path}{a path to a directory containing the exported .xlsx files}

\item{pattern}{a regex pattern for further selecing files in the directory;
defaults to "Tm Spec"}

\item{header}{if TRUE skips first 3 rows of .xlsx file to remove Uncle header; default is TRUE}

\item{combine}{if TRUE, returns all imported data merged into one unified dataframe with an "origin" column listing the original file path,
FALSE will return a list of dataframes; default is TRUE}
}
\value{
a named (with filename) list of dataframes or a single merged dataframe
}
\description{
\code{import_FLUORspec}
}
