% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLS.R
\name{import_SLSspec}
\alias{import_SLSspec}
\title{Import UNcle Tagg spectra into R}
\usage{
import_SLSspec(
  directory_path,
  pattern = "SLS Spec",
  lambda = 266,
  header = TRUE,
  combine = TRUE
)
}
\arguments{
\item{directory_path}{a path to a directory containing the exported .xlsx files}

\item{pattern}{a regex pattern for further selecing files in the directory;
defaults to SLS Spec}

\item{lambda}{a number value representing wavelength for Tagg spectra, typically 266nm for small aggregates and 473nm for large aggregates;
default is 266nm}

\item{header}{if TRUE skips first 1 rows of .xlsx file to remove UNcle header; default is TRUE}

\item{combine}{if TRUE, returns all imported data merged into one unified dataframe with an "origin" column listing the original file path,
FALSE will return a list of dataframes; default is TRUE}
}
\value{
a named (with filename) list of dataframes or a single merged dataframe
}
\description{
\code{import_SLSspec}
}
