% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_staticBundle}
\alias{import_staticBundle}
\title{Import static Uncle spectra (nanoDSF and SLS) into R}
\usage{
import_staticBundle(
  directory_path,
  pattern = "SLS Bundle",
  skip = 3,
  combine = TRUE
)
}
\arguments{
\item{directory_path}{a path to a directory containing the exported .xlsx files}

\item{pattern}{a regex for narrowing selection of files in the `directory_path`;
defaults to "SLS Bundle"}

\item{skip}{number of rows of embedded run metadata in the .xlsx file to remove; default is 3}

\item{combine}{if TRUE, returns all imported data merged into one unified dataframe with an "origin" column listing the original file path,
FALSE will return a list of dataframes; default is TRUE}
}
\value{
a named (with filename) list of dataframes or a single merged dataframe
}
\description{
\code{import_staticBundle}
}
