% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{join_SLS_DLS}
\alias{join_SLS_DLS}
\title{Joins imported and parsed DLS and Tm/Tagg SLS summary dataframes from the same experiment}
\usage{
join_SLS_DLS(SLS_data, DLS_data, .by = c("well", "sample_num",
  "prot_conc"))
}
\arguments{
\item{SLS_data}{a dataframe object containing SLS data}

\item{DLS_data}{a dataframe object containing DLS data}

\item{.by}{a character vector of variable names by which to join tables;
defaults to c("well", "sample_num", "prot_conc")}
}
\value{
a dataframe
}
\description{
\code{join_SLS_DLS}
}
