% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLS.R
\name{import_SLSsum}
\alias{import_SLSsum}
\title{Import UNcle Tm/Tagg SLS summary into R}
\usage{
import_SLSsum(directory_path, pattern = ".*\\\\.xlsx", sheet = NULL,
  header = FALSE)
}
\arguments{
\item{directory_path}{a path to a directory containing the exported .xlsx files}

\item{pattern}{a regex pattern for further selecing files in the directory,
defaults to reading all .xlsx files present}

\item{sheet}{character string to specify sheet if multi-sheet workbook is exported}

\item{header}{if TRUE skips first 4 rows of .xlsx file to remove UNcle header, default is FALSE}
}
\value{
generates a named list of tibbles,
each neamed element is one dataframe named with its origin file path
}
\description{
\code{import_SLSsum}
}
